+++ jestGlobalSetup.js
@@ -1,0 +1,6 @@
+/* eslint-disable import/no-extraneous-dependencies */
+const aggregateTranslations = require('terra-aggregate-translations');
+
+module.exports = () => {
+  aggregateTranslations();
+};