+++ packages/terra-application-name/tests/wdio/application-menu-name-spec.js
@@ -7,26 +7,13 @@ viewports.forEach((viewport) => {
     describe('Displays a default application menu name', () => {
       before(() => browser.url('/#/raw/tests/terra-application-name/application-name/application-menu-name-default'));
 
-      Terra.should.validateElement({ selector: '#default' });
-      if (browser.options.formFactor === 'medium') {
-        Terra.should.themeCombinationOfCustomProperties({
-          testName: 'themed',
-          properties: {
-            '--terra-application-name-menu-align-items': 'flex-start',
-            '--terra-application-name-menu-color': '#800080',
-            '--terra-application-name-menu-accessory-padding-right': '1rem',
-            '--terra-application-name-menu-title-font-weight': 'bold',
-            '--terra-application-name-menu-padding-left': '1rem',
-            '--terra-application-name-menu-padding-right': '3rem',
-          },
-        });
-      }
+      Terra.it.validatesElement({ selector: '#default' });
     });
 
     describe('Displays a wrapping application menu name', () => {
       before(() => browser.url('/#/raw/tests/terra-application-name/application-name/application-menu-name-wrapping'));
 
-      Terra.should.validateElement({ selector: '#wrapping' });
+      Terra.it.validatesElement({ selector: '#wrapping' });
     });
   });
 });