+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DisabledDatePickerField.test.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import DatePickerField from '../../../DatePickerField';
+
+const DatePickerDisabledDate = () => (
+  <DatePickerField
+    label="Enter Date"
+    name="date-input"
+    selectedDate="2017-04-01"
+    disabled
+  />
+);
+
+export default DatePickerDisabledDate;