+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Updated daylight savings time/standard time button to have aria-label and title with additional timezone info, to clarify its meaning.
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
 
 4.2.0 - (May 21, 2019)
 ------------------